+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - surfaces.Plane3D: linesegment_intersections
 - Step export
 - Face3D: is_linesegment_crossing.
+- BSplineCurve3D: from_step.
 - Step import
 - PeriodicalSurface: linesegment3d_to_2d, takes into account small 3D line segments that should be actually 3D arcs
 - babylondata: removes empty objects.
@@ -45,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - intersections: get_bsplinecurve_intersections generalization, so it can also be used
 to calculate intersections between a plane 3d and bsplinecurve3d.
 - Big refactor: New module curves.py containing classes as Line, Circle and Ellipse.
-Most edges will now be formed by a curve and start and end points. Unittests for all these classes have been created.
+Most edges will now be formed by a curve and a start and end points. Unittests for all these classes have been created.
 All adequations have been done for all tests and existing scripts.
 
 - bspline_compiled: refactor binomial_coefficient for performance.